cloned main branch

+++ include/png.h
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: okinnune <okinnune@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/19 17:12:25 by okinnune          #+#    #+#             */
-/*   Updated: 2022/10/19 18:15:53 by okinnune         ###   ########.fr       */
+/*   Created: 2022/07/08 11:11:21 by okinnune          #+#    #+#             */
+/*   Updated: 2022/10/20 20:08:30 by okinnune         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 typedef struct s_pngpalette
 {
-	uint64_t	*plte;
+	uint32_t	*plte;
 	uint32_t	length;
-} t_pngpalette;
+}	t_pngpalette;
 
 typedef struct s_pngdata
 {
